added comment feature

+++ app/routers/user.py
@@ -3,6 +3,7 @@ from app.utils import hash
 from app import models, schemas
 from sqlalchemy.orm import Session
 from app.database import get_db
+from app.oauth2 import get_current_user
 
 router = APIRouter(
     tags=['Users'],
@@ -28,7 +29,7 @@ def all_user(db:Session=Depends(get_db)):
 
 
 @router.get('/{id}', response_model=schemas.UserOut)
-def show(id:int, db:Session=Depends(get_db)):
+def show(id:int, db:Session=Depends(get_db), current_user:int= Depends(get_current_user)):
     user = db.query(models.User).filter(models.User.id==id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"No user available with id {id}")
@@ -36,10 +37,12 @@ def show(id:int, db:Session=Depends(get_db)):
 
 
 @router.delete('/{id}')
-def destroy(id:int, db:Session=Depends(get_db)):
+def destroy(id:int, db:Session=Depends(get_db), current_user:int= Depends(get_current_user)):
     user = db.query(models.User).filter(models.User.id==id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=f"No user available with id {id}")
+    if user.id == current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform actions")
     db.delete(user)
     db.commit()
     return "Deleted"
